+++ AGILA/Source/AGILA/i_am_a_fuckin_camera_PAWN.h
@@ -131,5 +131,5 @@ public:
 	UFUNCTION(BlueprintCallable)
 		void increaseMaxHealth(int increase);
 	UFUNCTION(BlueprintCallable)
-		void increaseSpeed(int increaseSpeed);
+		void increaseSpeed(float increaseSpeed);
 };
